requirements.txt
flask
flask-caching
flask-swagger-ui
gunicorn
python-dotenv
requests
redis
pandas
pyyaml
mstrio-py
openpyxl
